src/requirements.txt
flask
Pillow
requests
google-genai
numpy
opencv-python
pytesseract
pyttsx3
transformers
torch
python-dotenv
